+++ action_server/src/sema4ai/action_server/__init__.py
@@ -1,6 +1,6 @@
 from typing import List
 
-__version__ = "2.16.4"
+__version__ = "2.17.0"
 version_info = [int(x) for x in __version__.split(".")]
 
 __all__: List[str] = []